Abstracted download api logic
I mistakenly added the new download handler and route to the game api
module. Separated this logic into a separate download file.
